Make title case correction optional.
It's a good first pass but there's lots of cases where you want to make
manual corrections then reapply tag ordering and timestamp fixing.
Editing the script every time to make tag_titlecase() a no-op is no fun,
so I'm making it flaggable. And in keeping with the application of
ReplayGain, it's opt-in, not opt-out. I want the script to do magic, but
only the magic the user wants it to do.